Refactor: Move dtos to ee directory.
