+++ src/main/kotlin/com/wafflestudio/csereal/core/about/api/res/AboutSearchResBody.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.csereal.core.about.database
+package com.wafflestudio.csereal.core.about.api.res
 
 data class AboutSearchResBody(
     val total: Long,
